+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - HappyBirthday app which is one of the projects of [android-basics-compose-unit-1-pathway-3](https://developer.android.com/courses/pathways/android-basics-compose-unit-1-pathway-3)
+
+- ComposeArticle app which is one of the projects of [android-basics-compose-unit-1-pathway-3](https://developer.android.com/courses/pathways/android-basics-compose-unit-1-pathway-3)